networkd: Add settings for Calico's IP-in-IP tunnel
The tunnel interface tunl0 was getting the default setup of DHCP
and TX offload. Since Calico set up the IP address, the interface
was working but thought to be in "configuring" state by networkd
which is harmless but still not optimal. Recently the TX offload
caused Pod-to-Pod communication through the tunnel to break because
the checksums were not calculated.
Set the tunnel interface to be not managed by networkd's default
setup with DHCP. Disable the TX offload to enforce checksum
calculation.
See https://github.com/flatcar-linux/Flatcar/issues/183