Do not create empty /etc/coreos symlink
The symlink had no target and was only working on the final image after
the target folder got created by some other component. Now the target
folder and symlink is created by a tmpfile directive in baselayout and
the symlink here conflicts with the baselayout package's file.
Remove it since it is a duplicate.