network: Exclude new Azure driver as well
The bonded SR-IOV interface should not be configured individually and
has to be set to unmanaged. This is done by looking for	the driver but
now a new driver is in use and the list	has to be expanded or otherwise
the interface is stuck in a degraded state causing the
systemd-networkd-wait-online.service target to fail.