network: Restore KeepConfiguration=dhcp-on-stop
The default behavior in systemd-networkd was changed in v244 from
keeping the IP addresses and routes on service stop to deconfiguring
them:
https://github.com/systemd/systemd/commit/800603524a7ab076d40450286ff7802a04289b7f
Deconfiguring means that on system shutdown the DHCP address is
properly released but also has the side effect that orphaned processes
not part of a systemd unit don't have network connectivity when they
get the broadcasted SIGTERM.
Restore the previous behavior and hope that DHCP servers recognize
the system again on reboot and hand out the same address don't have to
rely on the address release (which, anyway, is not sent for a crashing
system either). The default can of course be changed by the user.
In the initramfs the KeepConfiguration=no behavior is desired because
otherwise the IP address is not released which can cause problems when
a different DHCP client configuration is set on first boot and the DHCP
server wouldn't be able to recognize the rootfs system and keep two
addresses allocated.

Fixes https://github.com/flatcar-linux/Flatcar/issues/213