extend-filesystems: More robustness against kernel partition re-reads
Directly using the device after the resize is prone to races.
At least for the size check it's easy to retry until we succeed getting
the new size. The filesystem resize call may still fail but the service
is limited in this way already because at startup it only checks if the
partition size is maximal but not the filesystem size. For future
debugging the output is made more verbose.

https://github.com/kinvolk/Flatcar/issues/235