networkd: avoid managing veth dev MAC addresses
When a veth device is created, the CNI in charge of bringing the device
up will set a MAC address, if `MACAddressPolicy=permanent` is set,
systemd will change it to a different one, causing dropped packets due
to mismatches.

With this change, the address set when the device is created will remain
untouched by systemd.

See https://github.com/kinvolk/Flatcar/issues/278 for more information.