bin/flatcar-install: add parameters to make wget more resilient
When the network has temporary issues it can happen that wget fails to
establish a connection or complete the transfer, or worse, get stuck
in a TCP connection which is not moving forward.
Add timeout and retry parameters to prevent unnecessary failure if a
retry fixes the problem and also to prevent the connection from getting
stuck.