Set hostname for EC2 and OpenStack from metadata
Similar to what is done for DigitalOcean and Packet, we can set the
hostname for EC2 and OpenStack through afterburn.
For OpenStack the name in afterburn does not match the oem_id set on
the kernel commandline from GRUB, and a special service is added
which sets the provider directly.