udev/rules.d: replace 79-net-google-compat.rules hardcoding ens4v1
With the systemd/udev 247 change of bind/unbind the old rule to set
ens4v1 as interface name on GCE broke and only eth0 was used now.
The old name ens4v1 is not even a current predictable name and the
breakage already appeared on Stable but for compatibility we can
support this name as alternative interface name instead of
maintaining the udev rule. For that, add special networkd link
files for GCE.