Avoid tearing down the network when using netroot
This change is based on the CoreOS support for Oracle OCI bare metal
machines, that was removed in 2018 with https://github.com/kinvolk/bootengine/commit/deba0732daec569545cf456f0cc514f17c7529b5.

It's been adapted to modern systemd versions (using KeepConfiguration
instead of CriticalConnection) and it works for any netroot, not just
Oracle, as we never want to tear down the network if we're using a
netroot, otherwise the root goes away when systemd does the switch-root
operation.