bin/flatcar-install: detect device mapper usage
The search for free disks was aware of mount points but did not
consider device mapper usage (e.g., LUKS or LVM) and went on to
overwrite the disk while it still was used. This can disrupt the
installation process or lead to unnoticed data corruption. It was also
not save when run on a system with active disk usage.
For disks with active device mapper entries the user must first remove
the mappings. For LVM this can be done with
lvchange -an /dev/mapper/PREFIX* (just /* also works to disable all LVM
mappings).

Fixes https://github.com/kinvolk/Flatcar/issues/332