network: don't manage the loopback interface
The default network configuration is DHCP and was also applied to the
loopback interface. This resulted in the loopback interface being in a
"managed" state in networkd and sometimes caused the IP address to be
deconfigured, resulting in errors when something tries to make use of
the loopback interface, like resolved
(see https://flaviutamas.com/2019/fixing-name-resolution-after-sleep).
Exclude the loopback interface from the default configuration to make
it "unmanaged".