dracut: add norecovery mount option for /usr
The /usr partition should not be modified during mounting, even if it
has some corruption. Rewriting the filesystem log would cause dm-verity
errors when dm-verity is enabled later again. While the /usr partition
normally is on a dm-verity block device in read-only mode there is some
option to mount the partition without dm-verity and it wouldn't be a
read-only block device anymore.
Add the norecovery mount option which is supported for ext4 and btrfs.