dracut: fall back to expected dm-verity hash offset
The hash offset is found by looking at the filesystem size. When
e2size can't find the size it returns "Success" in stderr for whatever
reason and fortunately still returns an error exit code, stdout stays
empty. This means that the dm-verity device setup won't work because
the hash offset is the empty string. However, the hash offset is
actually fixed because the GPT disk layout has to stay the same in
Flatcar Container Linux as the partition contents are swapped out
when updating. In case another filesystem like btrfs is used, e2size
doesn't work and it makes sense to fall back to the only value which
is supported in general.
Hard code the hash offset value coming from the /usr filesystem size
defined in flatcar-scripts/build_library/disk_layout.json.