dracut: issue a kernel panic on dm-verity corruption
In the default mode dm-verity leads to failed read syscalls when
corruption is detected. This is hard to spot and also not secure
because it allows altering system behavior by introducing failures
to particular parts of the system and thus turning off security
mechanisms. Thus, the value of dm-verity isn't really utilized.

Issue a kernel panic instead of just failing a read syscall when
corruption is detected. This allows users to spot the issue and
take action and fully utilizes the advantages of having dm-verity.