bin/flatcar-install: randomize OEM filesystem UUID if mounting fails
When an identical BTRFS filesystem is already mounted because the
installing system has the same OEM partition version, the mounting
of the installed OEM partition fails because identical UUIDs are
only allowed to be used for btrfs multi device setups.
Randomize the filesystem UUID in case mounting fails and try again.