dracut: run emergency.target on ignition/torcx service unit failure
Right now, when ignition fails during boot the following happens:

1. ignition service fails, which fails initrd.target fails and starts emergency.target
2. initrd-parse-etc.service is already starting and launches initrd-fs.target and initrd-cleanup.service
3. initrd-cleanup.service isolates initrd-swithc-root.target
4. emergency.service is stopped and starts default.target again
5. initrd.target starts ignition services again
6. loop back to 1.

This is a boot loop that makes it hard for users to figure out what is wrong
with their ignition config. This appears to be unintentional, and something
observed in CoreOS and Fedora CoreOS as well.

Apparently, for a brief while systemd did not allow OnFailure clauses in
targets. This was fixed in systemd, but coreos shipped this fix as well
https://github.com/coreos/ignition-dracut/pull/188.

So add OnFailure clauses to ignition units to ensure boot stops right there.

Signed-off-by: Jeremi Piotrowski <jpiotrowski@microsoft.com>