Remove logrotate dependency from init
As the package app-admin/logrotate is moving from portage-stable
to coreos-overlay. We are moving the service file and the config
files to the package files directory in the coreos-overlay. So we
no longer need them in this repo. Also we are removing the timer
file and the timer symlink as we are using the upstream timer. And
we are removing the logrotate dependency from the Makefile.