bin: add flatcar-update tool for manual updates or airgapped updates
While updates are managed through Nebraska, not every user wants to or
can run their own service. Sometimes it would be more convenient if a
local update command exists that also supports airgapped updates or
helps a user to quickly jump between versions to reproduce an error.

Add a flatcar-update tool (similar to flatcar-install) which users can
run on the machine to force an update to a particular release or local
update payload. On old machines this script can be fetched and
executed, too, and should be CoreOS-compatible. Since update-engine
does no version check, it's also possible to downgrade with the tool.
The tool can disable updates after it ran, making it easy to do manual
updates when no own Nebraska server exists.