networkd: do not use DHCP for special devices
We are always lagging behind with which interface names are used by
CNIs and have to be set Unmanaged so that we don't use DHCP for them.

Go the other way round and exclude special devices which won't use DHCP
upfront so that we don't even try configuring them this way. Bridges
normally don't get their IP address assinged via DHCP and if someone
wants that, they are free to ship their own networkd unit. Same for
tunnel devices and similar. If this causes problems, we can revert this
because we still have the explicit rules but eventually this should
allow us to retire some of our rules that match by interface name.