dracut/99switch-root: use init.wrapper when cgroupv1 is requested
Pass a custom init commands (init.wrapper) to 'systemctl switch-root'
when leaving the initramfs. This init.wrapper is responsible for
disabling all cgroup controllers and unmounting cgroup2, and injecting
(by means of a bind mount) custom kernel arguments so that sysroot
systemd sets up cgroupv1. We're adding this so that Flatcar users don't
have to reboot to switch back to cgroupv1, which is disruptive in many
deployments.

The usage of init.wrapper is conditional on two things:
* initramfs used up cgroupv2.
* user opted into to this behavior by creating the
  `/etc/flatcar-cgroupv1` flag file

There is a reason to not go through init.wrapper unconditionally, and
decide whether to take action-or-not there. When systemd runs in sysroot
and initramfs, it serializes some data structures and passes them from
the initramfs to allow some introspection. I believe this is primarily
systemd-analyze timing information. This doesn't work when we inject a
custom init between sysroot and initramfs (init.wrapper), so ensure we
only use the custom init when we really need to.

This is accomplished by passing the init.wrapper path through an INIT
environment variable, which is only defined when '/etc/flatcar-cgroupv1'
exists in the sysroot. An alternative would have been to conditionally
symlink the override snippet, which would get applied in the 'systemctl
daemon-reload' in 'initrd-parse-etc.service'. Neither option is
particularly obvious, but this one seems slightly easier to follow.