systemd: disable foreign route management
While systemd-networkd follows the principle of a declarative network
configuration and thus needs a way to ensure that unwanted routes or
routing policy rules get discarded, the interfacing with procedural
network management from CNIs like Cilium is limited, so that when the
interface is set to "unmanaged" through a networkd unit, any routing
policies there would also be ignored and discarded unless they would
be defined for a new unit for a dummy network interface. This means
the only option left is to disable the discarding of foreign rules
globally.

Set the default for ManageForeignRoutes and
ManageForeignRoutingPolicyRules to "no" to ensure that we don't
interfere with the network management of the CNIs. Users that rely on
the setting can still enable it again but only through a drop-in
under /etc/systemd/networkd.conf.d/ because this here is a drop-in
already that takes precedence over the top config file.

See https://github.com/cilium/cilium/issues/18706
and https://github.com/flatcar-linux/Flatcar/issues/620
Replaces https://github.com/flatcar-linux/coreos-overlay/pull/1622