systemd: add helper service to start sysext services
As systemd-sysext images are not really meant for system services
(that is the role of portable services) the services that they define
don't get started even if they include the symlinks to do so.
Add a helper service that starts the new services by retriggering the
socket, timers and multi-user targets as these are the targets a user
would normally use to enable units. We can add more as needed.