systemd/system/ensure-sysext: skip unit if systemd-sysext is skipped
The unit failed to execute when the dependency systemd-sysext.service
was skipped because it's not needed. Normally this is harmless but it
still triggers a OnFailure event as reported in
https://github.com/flatcar-linux/Flatcar/issues/710

Use the same logic for skipping execution as done in
systemd-sysext.service - we can't use "systemctl is-enabled" because
Flatcar's inbuilt unit symlinks are not reported as "enabled".