If the directory symlink exists, don't create a symlink in it
When motdgen ran again, we ended up with a "flatcar -> flatcar" symlink
inside the /run/flatcar/ directory because the existing /run/coreos
symlink is pointing to the /run/flatcar directory and without -T ln
will then create the symlink inside the given path instead of
overwriting the symlink.
Add the -T flag and also use it in the Makefile whenenver we want to
set up symlinks for directories.