flatcar-reset: Add backup and preview
When users run flatcar-reset from the command line to try it out, this
should not result in data loss when they forget to quote the keep paths
correctly or didn't know how to stop the staged reset.
Add a warning and the option to see the list of files in preview. Also
add the option to do a backup when staging a reset. The backup does not
cover ACLs because Flatcar's rsync isn't compiled with support for
them.