*: Prepare everything to use a new OEM mountpoint location
The OEM partition mountpoint will be moved to /oem. The scripts using
the old location needed an update. The install script still check the
old location, just in case, but checks the new location first. The new
mountpoint location also means that some systemd unit names changed,
so we need to adapt. Also, this drops a sysext workaround that is not
necessary any more.