flatcar-install: Wait for subproccesses and report errors
The redirection of FD 3 to the input FD with >(write_to_disk) causes
the function to not wait or react on errors from write_to_disk.
Before returning, wait for all subprocesses to finsh and then check the
success through the out-of-band FIFO. Since we don't need the waiting
with a large timeout anymore we can change the wait function to fulfill
a new purpose.