flatcar-update: Support Flatcar OEM and extension payloads
The OEMs are now getting ported over to systemd-sysext images and they
are delivered as additional update payloads in the Omaha response. We
also define optional Flatcar extensions that the user can enable. While
update-engine's post-install action and the initrd have a fallback
mechanism that use the release server in case flatcar-update does not
provide the required payloads, this does not work for airgapped
environments or updating to developer payloads.
Let flatcar-update download the required payloads for the running
machine from the release server instead of relying on any fallback logic
and also request the user to provide any required extension payloads.