systemd/network: Disable RoutesToDNS
Flatcar VMs on Azure with multiple nics fail to provision correctly because of
multiple routes to wireserver (168.63.129.16) with the same metric, and
wireserver only responds to 'ready' signals on the primary nic. The primary nic
gets an explicit route to wireserver through dhcp. It turns out systemd is
adding the extra routes because wireserver is also the dhcp and dns server, and
the default RoutesToDNS=true setting is the reason.

This behavior is non-standard and I don't think it's expected on any platform.
Disable this behavior by default.

Signed-off-by: Jeremi Piotrowski <jpiotrowski@microsoft.com>