systemd/system/sshkeys: Make execution more robust
The service failed when the unit to start was masked. Since it involves
networking, the service can fail and if it does we should restart. It
also seems that the service relied on
update-ssh-keys-after-ignition.service to run update-ssh-keys once to
create the subfolder and could possibly race with it.

Prevent execution when it has no chance to succeed because the unit to
start is masked but in all other cases try to restart on failure. Also
order behind update-ssh-keys-after-ignition.service.