flatcar-update: Support flag to skip providing OEM payloads
The required use of OEM payloads can be problematic when trying to
downgrade, or for testing the fallback case.
Add a flag to skip providing OEM payloads. This also allows us to
backport the support for OEM payloads to LTS, by switching the default
behavior of the flag to "no".