extend-filesystems: Fix race condition by not using lsblk
lsblk relies on udev, which is inherently racy, especially when the
filesystem has just been mounted. It has been observed that the FSTYPE
field is sometimes not populated, preventing the filesystem from being
resized, and causing tests involving the dev container (which is large)
to fail.

Use findmnt instead, which gets its information directly from the
kernel. It also natively supports filtering by filesystem type and mount
option. It doesn't fetch PARTTYPE, but we can safely get that from cgpt.

Signed-off-by: James Le Cuirot <jlecuirot@microsoft.com>