Make systemd-confext automatically use the mutable mode
So far we had a custom overlay mount for /etc that provided the A/B
updated files from /usr in a lowerdir. Since then we upstreamed a
mutable mode for sysext and confext.
We can now switch over to it and provide a default confext by using the
mutable mode. For user-provided to be supported we need to wait for the
atomic remount to be there so that a reload is less impacting. Also, one
has to be careful with the stacking order for user-provided extensions
and modes other than mutable. The read-only mode also needs more design
work in Flatcar.

Signed-off-by: Kai Lueke <kailuke@microsoft.com>