Exclude TUN devices from default systemd-networkd setup
The default is to configure a network interface with systemd-networkd
but we exclude those devices where it doesn't make sense. This was not
done for TUN devices and we should exclude them as well, assuming that
the creator would also configure it.
Instead of matching against the "none" type TUN interfaces have, use the
driver and kind property to exclude TUN interfaces and a few more while
at it (but they are already excluded and we add them just for
completeness).

Signed-off-by: Kai Lueke <kailuke@microsoft.com>