Drop old ensure-sysext.service workaround
When systemd-sysext didn't support issuing a daemon reload, we had this
service do it. This is now doing the same thing again while not needed.
The service also added support to using .wants but in general .upholds
should be used. For extensions that use .wants we can still start their
services when we set the extension up from the initrd. Left is just the
case where the extension is loaded live but since this was a
Flatcar-specific workaround it's not something we want to support and
users either should migrate to .upholds, start the service manually (or
with the target restart as done in this workaround), or do a reboot when
they add new extensions live. We also plan to add a service restart
field to the extension-release metadata file so that one doesn't need to
use .upholds and it even covers more cases.

Signed-off-by: Kai Lueke <kailuke@microsoft.com>