systemd: Define order for confext and sysext
For a confext to influence how things shipped in a sysext behave or even
how sysext itself behaves it's good to define that confext runs first.
This is also explicitly done in bootengine. With the skip logic we
ideally don't set up extensions again during boot but that only works if
the contents can be ensured to be equal and the fallback is a full
refresh otherwise.

Signed-off-by: Kai Lueke <kailuke@microsoft.com>