systemd: Add noop sysupdate transfer config
When the systemd-sysupdate.service runs it fails if there is no config.
Since we don't use it, it's expected to do nothing but this failure is
making problems when the service is started from the timer unit which
is now enabled by default. This service is also used by the
sysext-bakery update configs for drop-in steps that update sysupdate
components. There we let users add a noop config to prevent this service
failure.
Since we run into this in Flatcar now and we anyway had users add this
manually, let's just add it directly to Flatcar.

Signed-off-by: Kai Lueke <kailuke@microsoft.com>